build

+++ src/include/OrderHistory.hpp
@@ -2,6 +2,8 @@
 
 #include "Structure.hpp"
 
+#include <cstdint>
+
 #define ORDER_HISTORY_POPUP_WINDOW "Order History Window"
 class OrderHistory : public Singleton {
   public:
@@ -17,6 +19,6 @@ class OrderHistory : public Singleton {
   private:
     BookOrderListT _container;
     bool           _showOrderHistory = true;
-    double         _orderNumber      = 0;
+    uint64_t       _orderNumber      = 0;
     int            _index            = 0;
 };
